feat: Added actionable insights for telegram bot

+++ src/main/java/com/hiringassignment/hiringassignment/entity/RedditTopPost.java
@@ -42,7 +42,7 @@ public class RedditTopPost {
     private String permalink;
 
     @Column(name = "num_comments")
-    private Integer numComments;
+    private Long numComments;
 
     @Column(name = "created_at")
     private LocalDateTime createdAt;
